Move webpack files and rename
This cleans up the root project folder a bit and makes it easier to figure out
the function of each webpack file.
