+++ package.json
@@ -12,18 +12,18 @@
     "lint": "./node_modules/.bin/eslint app --ext .js --ext .jsx && ./node_modules/.bin/eslint test --ext .js && ./node_modules/.bin/stylelint 'app/**/*.css'",
     "lint-fix": "./node_modules/.bin/eslint app --ext .js --ext .jsx --fix && ./node_modules/.bin/eslint test --ext .js --fix && ./node_modules/.bin/stylelint 'app/**/*.css' --fix",
     "hot-server": "cross-env NODE_ENV=development node --max_old_space_size=4096 -r @babel/register server.js",
-    "build-main": "cross-env NODE_ENV=production node -r @babel/register ./node_modules/webpack/bin/webpack --config webpack.config.electron.production.js --progress=profile --color",
-    "build-renderer": "cross-env NODE_ENV=production node -r @babel/register ./node_modules/webpack/bin/webpack --config webpack.config.production.js --progress=profile --color",
-    "build-trezor": "cross-env NODE_ENV=production node -r @babel/register ./node_modules/webpack/bin/webpack --config webpack.config.trezor.js --progress=profile --color",
-    "build-preload": "cross-env NODE_ENV=production node -r @babel/register ./node_modules/webpack/bin/webpack --config webpack.config.preload.js --progress=profile --color",
+    "build-main": "cross-env NODE_ENV=production node -r @babel/register ./node_modules/webpack/bin/webpack --config webpack/electron.prod.js --progress=profile --color",
+    "build-renderer": "cross-env NODE_ENV=production node -r @babel/register ./node_modules/webpack/bin/webpack --config webpack/ui.prod.js --progress=profile --color",
+    "build-trezor": "cross-env NODE_ENV=production node -r @babel/register ./node_modules/webpack/bin/webpack --config webpack/trezor.js --progress=profile --color",
+    "build-preload": "cross-env NODE_ENV=production node -r @babel/register ./node_modules/webpack/bin/webpack --config webpack/preload.js --progress=profile --color",
     "build": "npm run build-trezor && npm run build-preload &&  npm run build-main && npm run build-renderer",
     "dexsite": "node ./scripts/dexsite.js",
     "rebuild-natives": "node_modules/.bin/electron-rebuild",
     "rebuild-dexc": "cd modules/dex && npm run install",
     "start": "cross-env NODE_ENV=production electron ./app/ --debug --custombinpath=./bin",
-    "start-hot": "webpack --progress -c webpack.config.electron.development.js && cross-env HOT=1 NODE_ENV=development electron -r @babel/register -r @babel/polyfill ./app/main.js",
-    "start-hot-nosandbox": "webpack -c webpack.config.electron.development.js &&  cross-env HOT=1 NODE_ENV=development electron -r @babel/register -r @babel/polyfill ./app/main.js -r --no-sandbox",
-    "start-preload": "webpack -c webpack.config.preload.js --watch",
+    "start-hot": "webpack --progress -c webpack/electron.dev.js && cross-env HOT=1 NODE_ENV=development electron -r @babel/register -r @babel/polyfill ./app/main.js",
+    "start-hot-nosandbox": "webpack -c webpack/electron.dev.js &&  cross-env HOT=1 NODE_ENV=development electron -r @babel/register -r @babel/polyfill ./app/main.js -r --no-sandbox",
+    "start-preload": "webpack -c webpack/preload.js --watch",
     "postinstall": "electron-builder install-app-deps && npm run rebuild-dexc && npm rum dexsite",
     "dev": "npm run build-preload && npm run hot-server -- --start-hot",
     "dev-nosandbox": "npm run hot-server -- --start-hot-nosandbox",
