+++ server.js
@@ -11,7 +11,7 @@ import webpackDevMiddleware from "webpack-dev-middleware";
 import webpackHotMiddleware from "webpack-hot-middleware";
 import { spawn } from "child_process";
 
-import config from "./webpack.config.development";
+import config from "./webpack/ui.dev.js";
 
 const argv = require("minimist")(process.argv.slice(2));
