+++ webpack/ui.base.js
@@ -56,7 +56,7 @@ export default {
   },
 
   output: {
-    path: path.join(__dirname, "app"),
+    path: path.join(__dirname, "../app"),
     filename: "bundle.js",
 
     // https://github.com/webpack/webpack/issues/1114
@@ -71,12 +71,12 @@ export default {
     extensions: [ ".js", ".jsx", ".json" ],
     mainFields: [ "webpack", "browser", "web", "browserify", [ "jam", "main" ], "main" ],
     alias: {
-      fetchModule: path.resolve(__dirname, "app/helpers/fetchModule.js"),
-      walletCrypto: path.resolve(__dirname, "app/helpers/walletCryptoModule.js")
+      fetchModule: path.resolve(__dirname, "../app/helpers/fetchModule.js"),
+      walletCrypto: path.resolve(__dirname, "../app/helpers/walletCryptoModule.js")
     },
     modules: [
-      path.resolve(__dirname, "app"),
-      path.resolve(__dirname, "app/components"),
+      path.resolve(__dirname, "../app"),
+      path.resolve(__dirname, "../app/components"),
       "node_modules"
     ]
   },
