+++ webpack/ui.dev.js
@@ -5,7 +5,7 @@
 
 import webpack from "webpack";
 import merge from "webpack-merge";
-import baseConfig from "./webpack.config.base";
+import baseConfig from "./ui.base";
 import HtmlWebpackPlugin from "html-webpack-plugin";
 import CopyWebpackPlugin from "copy-webpack-plugin";
